requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
drf_spectacular==0.27.0
pytz==2023.3.post1
psycopg2-binary
